chore(e2e): split up tests and rename
- to be captured by ci tests need to start with e2e
- each e2e test file should have a single test
